+++ providers/google/src/airflow/providers/google/cloud/operators/dataproc/__init__.py
@@ -1,0 +1,35 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+"""Google Dataproc operators package.
+
+DataprocCreateBatchOperator is lazy-loaded from ._batch to avoid pulling in
+heavy google.cloud.dataproc_v1 and hook dependencies at DAG parse time (#62373).
+All other operators are loaded from ._core on first access.
+"""
+
+from __future__ import annotations
+
+
+def __getattr__(name: str):
+    if name == "DataprocCreateBatchOperator":
+        from airflow.providers.google.cloud.operators.dataproc._batch import (
+            DataprocCreateBatchOperator,
+        )
+        return DataprocCreateBatchOperator
+    from airflow.providers.google.cloud.operators.dataproc import _core
+    return getattr(_core, name)
